problem comes from BaseMapping doesn't have implemented call method, so we're trying to implement it
